rename(generator): generator 패키지 형성
