+++ docs/README.md
@@ -145,7 +145,8 @@
      - [x] 정렬된 로또 번호를 저장한다
    - 저장된 로또 번호를 조회한다
      - [x] 인덱스에 따라 로또 번호를 조회할 수 있다 
-
+     - [x] 인덱스에 저장된 로또 객체를 조회한다
+     - [x] 로또 객체가 가지고 있는 List<Integer>을 반환한다 
 5. LottoGame에서 사용자의 로또 번호와 생성된 로또 번호를 비교한다 
 
 6. 당첨 통계를 확인한다
