+++ src/test/java/lotto/LottoRepositoryTest.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.stream.Stream;
 import lotto.domain.Lotto;
 import lotto.domain.LottoGame;
+import lotto.generator.MockedRandomGenerator;
 import lotto.repository.LottoRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
